+++ src/main/java/ru/practicum/shareit/item/service/ItemService.java
@@ -5,16 +5,16 @@ import ru.practicum.shareit.item.model.ItemDto;
 import java.util.List;
 
 public interface ItemService {
-    ItemDto create(Integer userId, ItemDto itemDto);
+    ItemDto create(Long userId, ItemDto itemDto);
 
-    ItemDto get(Integer id);
+    ItemDto get(Long id);
 
-    List<ItemDto> getAllItemUsers(Integer userId);
+    List<ItemDto> getAllItemUsers(Long userId);
 
     List<ItemDto> search(String text);
 
-    ItemDto update(Integer userId, Integer itemId, ItemDto itemDto);
+    ItemDto update(Long userId, Long itemId, ItemDto itemDto);
 
-    void delete(Integer id);
+    void delete(Long id);
 
 }
